Create Shapesarea.java
Abstraction is the process of hiding certain data and showing only the necessary details. Abstract method has to be created using abstract keyword and we cannot create instance for it. This can be achieved through method overriding. Here, I have written a program for finding area of two shapes where the area is abstract method and shapes like Circle and Rectangle are the other different classes where area is found through overriding.